pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hackathon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.118.0",
    "google-adk>=1.15.1",
    "google-auth>=2.41.1",
    "google-cloud-firestore>=2.21.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
    "pypdf>= 4.0.0",
    "google-generativeai>=0.7.2",
    "python-dotenv>=1.0.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]

[tool.uv]
dev-dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=6.30.1",
    "tqdm>=4.67.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
dev  = "src.app.main:dev_server"
prod = "src.app.main:prod_server"
